+++ CHANGELOG.md
@@ -5,3 +5,6 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 ## 0.0.1
 
 - Initial commit of extension that provides simple menu to call `kamel run --dev "selectedfile"`
+- Added PR from Zoran adding a 'Stop Camel-K integration' menu
+- Added new "Camel-K Integrations" view with refresh (view action) and remove (context action)
+- Improved other actions so view is refreshed when integrations are started and stopped